% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_sem.R
\name{df_sem}
\alias{df_sem}
\title{Generate standard error of mean}
\usage{
df_sem(data, x)
}
\arguments{
\item{data}{A data frame, generally grouped by the intended sub-groups which you want to compare for the
same t-test.}

\item{x}{A (bare) column name of the variable which you want to get the mean, sd, and standard error of the mean (SEM).}
}
\value{
A \code{data.frame} with
consisting of characters. The columns that are always present are:
\code{group variable(s)}, \code{mean}, \code{sd}, \code{n}, and \code{se}(SEM).
}
\description{
Generate standard error of mean
}
\examples{
df_sem(color_index, color_index)

library(magrittr)
color_index \%>\%
  df_sem(color_index)
}
