% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roistats-package.R
\docType{package}
\name{roistats-package}
\alias{roistats}
\alias{roistats-package}
\title{roistats: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Yufei Zhao \email{fay.zz0213@gmail.com} (\href{https://orcid.org/0000-0002-6511-8507}{ORCID})

}
\keyword{internal}
