% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roistats-package.R
\docType{package}
\name{roistats-package}
\alias{roistats}
\alias{roistats-package}
\title{roistats: Convienience functions for applying basic stats to sub-groups}
\description{
The goal of this package is for easily applying same t-tests/basic data description across several sub-groups, with the output as a nice arranged data.frame. Multiple comparison and the significance symbols are also provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Irisfee/roistats}
  \item Report bugs at \url{https://github.com/Irisfee/roistats/issues}
}

}
\author{
\strong{Maintainer}: Yufei Zhao \email{fay.zz0213@gmail.com} (\href{https://orcid.org/0000-0002-6511-8507}{ORCID})

}
\keyword{internal}
